Move LoadingSkeleton to components/ui directory
- Relocate LoadingSkeleton.tsx from components/ to components/ui/
- Update import path in ProposalList.tsx
- Remove unnecessary inline comments
